Fixes for events OnError and OnTick are now hooked up and event function returns use a specific return which allows them to be differentated from function returns so that calls to functions no longer interact with calls to events.

+++ Source/DslValue.cpp
@@ -1309,7 +1309,7 @@ void DslValue::printItem(bool showEscapes)
             printf("\n");
             return;
         case INTEGER_VALUE:
-            printf("%lld", iValue);
+            printf("%lld", (long long int)iValue);
             break;
         case DOUBLE_VALUE:
             printf("%f", dValue);
